ngrok solution

+++ lib/network.py
@@ -10,6 +10,7 @@ import urllib
 import uuid
 import os 
 import re 
+from os.path import expanduser
 
 BLUE, RED, WHITE, CYAN, DEFAULT, YELLOW, MAGENTA, GREEN, END, BOLD = '\33[94m', '\033[91m', '\33[97m', '\033[36m', '\033[0m', '\33[93m', '\033[1;35m', '\033[1;32m', '\033[0m', '\033[1m'
 
@@ -39,21 +40,25 @@ def public_ip():
 print("\n{0}Public IP: {1}{2}").format(GREEN, DEFAULT, public_ip())
 
 def ngrok():
-    a = os.popen('pgrep ngrok').read()
-    if not a:
-        os.system('./ngrok tcp 443 > /dev/null &')
-        while True:
-            os.system('curl -s -N http://127.0.0.1:4040/status | grep "tcp://0.tcp.ngrok.io:[0-9]*" -oh > ngrok.tcp')
-            TcpFile = open('ngrok.tcp', 'r')
-            tcp = TcpFile.read()
-            TcpFile.close()
-            if re.match("tcp://0.tcp.ngrok.io:[0-9]*", tcp) != None:
-                print("\n{0}Ngrok TCP: {1}{2}".format(GREEN, DEFAULT, tcp))
-                break
+    home = expanduser("~")
+    if os.path.isfile('{}/.ngrok2/ngrok.yml'.format(home)):
+        a = os.popen('pgrep ngrok').read()
+        if not a:
+            os.system('./ngrok tcp 443 > /dev/null &')
+            while True:
+                os.system('curl -s -N http://127.0.0.1:4040/status | grep "tcp://0.tcp.ngrok.io:[0-9]*" -oh > ngrok.tcp')
+                TcpFile = open('ngrok.tcp', 'r')
+                tcp = TcpFile.read()
+                TcpFile.close()
+                if re.match("tcp://0.tcp.ngrok.io:[0-9]*", tcp) != None:
+                    print("\n{0}Ngrok TCP: {1}{2}".format(GREEN, DEFAULT, tcp))
+                    break
 
+        else:
+            os.system('kill -9 $(pgrep ngrok)')
+            ngrok()
     else:
-        os.system('kill -9 $(pgrep ngrok)')
-        ngrok()
+        print("\n{0}Ngrok TCP:{1} None\n".format(GREEN, DEFAULT))
 
 ngrok()
